extracted dark mode toggle out to individual component
